+++ src/func/auth/login.js
@@ -1,5 +1,6 @@
 import jwt from 'jsonwebtoken';
 import moment from 'moment';
+import { intersection } from 'lodash';
 import { isMatchingWithHashedPassword } from './helper';
 import { NotFoundError, BadRequestError } from '../../lib/errors';
 
@@ -48,21 +49,17 @@ export async function getUserByEmail(req) {
 
   return {
     ...req,
-    user,
+    user: user.Content,
   };
 }
 
 export function verifyPassword(req) {
   const {
     instrumentation,
-    user: {
-      Content: {
-        password,
-      },
-    },
+    user,
   } = req;
 
-  if (!isMatchingWithHashedPassword(req.body.data.password, password)) {
+  if (!isMatchingWithHashedPassword(req.body.data.password, user.password)) {
     instrumentation.error('Password does not match');
 
     throw new BadRequestError('Password does not match.');
@@ -72,10 +69,36 @@ export function verifyPassword(req) {
 }
 
 export function generateToken(req) {
-  const expiresIn = parseInt('1800', 10); // ~30mins
-  const token = jwt.sign(req.user, process.env.APP_KEY, {
-    expiresIn, // in seconds
-  });
+  const {
+    user,
+    body: { data: { scopes } },
+  } = req;
+
+  // validate requested scopes
+  const grantedScopes = Array.isArray(user.scopes) ? user.scopes : user.scopes.split(' ');
+  const wantedScopes = Array.isArray(scopes) ? scopes : scopes.split(' ');
+
+  if (wantedScopes.length > grantedScopes.length) {
+    throw new BadRequestError('Your scopes you requested is not available.');
+  }
+  if (intersection(grantedScopes, wantedScopes) < wantedScopes.length) {
+    throw new BadRequestError('User does not have scopes.');
+  }
+
+  // generate token
+  const token = jwt.sign(
+    {
+      userEmail: user.email,
+      sub: user.email,
+      iss: 'api.heligram.com',
+      scopes: wantedScopes,
+    },
+    process.env.APP_KEY,
+    {
+      expiresIn: '30m',
+    },
+  );
+
   const expireAt = moment().add(30, 'm').unix();
 
   return {
@@ -92,7 +115,6 @@ export function returnResponse(req) {
       data: {
         type: 'session',
         attributes: {
-          email: req.body.data.email,
           token: req.token,
           expireAt: req.expireAt,
         },