+++ src/lib/middlewares/index.js
@@ -1,5 +1,7 @@
 import malformedErrorHandler from './malformedErrorHandler';
+import jwtAuthz from './jwtAuthz';
 
 export {
   malformedErrorHandler,
+  jwtAuthz,
 };