+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- Added `scopes` in JWT claims to handle API Authorization.
+- Added `ForbiddenError` to handle API Authorization Error.
+
 ### Changed
+
+- Changed `login-user+v1.json` schema to support `scopes` property in request payload
+
+### Fixed
+
+- Fixed wrong-working `jwtPassport` middleware.
 
 ### Removed